requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb>=0.4.15
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
python-dotenv>=1.0.0
openai>=1.6.0

